requirements.txt
fastapi==0.103.2
redis_om==0.3.2
python-dotenv==1.0.1
httpx==0.27.0